Add files via upload
This simple addon was specifically made for a project in which i needed to make sure that a loop of circles where to be connected in a chain such that each circle is tangent to the next. PS. this is also my first attempt at making a blender addon so proceed with caution.